feat: split ol_concourse.lib into standalone ol-concourse package
- Remove src/ol_concourse/lib/ (moved to mitodl/ol-concourse on PyPI)
- Remove src/ol_concourse/__init__.py (namespace package: ol-concourse + ol-infrastructure contribute to same namespace)
- Add ol-concourse git dependency (subdirectory=pipeline_lib) until 0.6.0 published to PyPI
- Add src/ol_concourse/pipelines/jobs.py: functools.partial wrapper for pulumi_jobs_chain pre-filled with GH_ISSUES_DEFAULT_REPOSITORY
- Update all pipeline imports to use new jobs.py wrapper instead of lib.jobs.infrastructure
- Update all meta-pipeline watched paths from 'src/ol_concourse/lib/' to 'pyproject.toml'
- Add INP001 ruff exception for src/ol_concourse/** (namespace package has no __init__.py)
- Regenerate uv.lock with resolved ol-concourse 0.6.0
