--- src/ol_concourse/lib/constants.py
@@ -1,1 +1,0 @@
-REGISTRY_IMAGE = "registry-image"
