+++ src/ol_concourse/pipelines/infrastructure/consul/packer_pulumi_pipeline.py
@@ -1,6 +1,6 @@
 import itertools
 
-from ol_concourse.lib.jobs.infrastructure import packer_jobs, pulumi_jobs_chain
+from ol_concourse.pipelines.jobs import packer_jobs, pulumi_jobs_chain
 from ol_concourse.lib.models.fragment import PipelineFragment
 from ol_concourse.lib.models.pipeline import GetStep, Identifier, Pipeline
 from ol_concourse.lib.resource_types import hashicorp_resource
