+++ src/ol_concourse/pipelines/infrastructure/dagster/docker_pulumi_pipeline.py
@@ -1,7 +1,7 @@
 import sys
 
 from ol_concourse.lib.containers import container_build_task
-from ol_concourse.lib.jobs.infrastructure import pulumi_jobs_chain
+from ol_concourse.pipelines.jobs import pulumi_jobs_chain
 from ol_concourse.lib.models.fragment import PipelineFragment
 from ol_concourse.lib.models.pipeline import (
     AnonymousResource,
