+++ src/ol_concourse/pipelines/infrastructure/k8s_apps/docker_pulumi.py
@@ -8,7 +8,7 @@ from pydantic import BaseModel, model_validator
 
 from ol_concourse.lib.constants import REGISTRY_IMAGE
 from ol_concourse.lib.containers import container_build_task
-from ol_concourse.lib.jobs.infrastructure import pulumi_job, pulumi_jobs_chain
+from ol_concourse.pipelines.jobs import pulumi_job, pulumi_jobs_chain
 from ol_concourse.lib.models.fragment import PipelineFragment
 from ol_concourse.lib.models.pipeline import (
     AnonymousResource,
