+++ src/ol_concourse/pipelines/infrastructure/keycloak/docker_pulumi_pipeline.py
@@ -4,7 +4,7 @@ import textwrap
 from bridge.lib.versions import KEYCLOAK_VERSION
 from ol_concourse.lib.constants import REGISTRY_IMAGE
 from ol_concourse.lib.containers import container_build_task
-from ol_concourse.lib.jobs.infrastructure import pulumi_jobs_chain
+from ol_concourse.pipelines.jobs import pulumi_jobs_chain
 from ol_concourse.lib.models.fragment import PipelineFragment
 from ol_concourse.lib.models.pipeline import (
     AnonymousResource,
