+++ src/ol_concourse/pipelines/infrastructure/kubewatch/pulumi_pipeline.py
@@ -3,7 +3,7 @@
 import sys
 
 from ol_concourse.lib.containers import container_build_task
-from ol_concourse.lib.jobs.infrastructure import pulumi_jobs_chain
+from ol_concourse.pipelines.jobs import pulumi_jobs_chain
 from ol_concourse.lib.models.fragment import PipelineFragment
 from ol_concourse.lib.models.pipeline import (
     GetStep,
