+++ src/ol_concourse/pipelines/open_edx/edx_platform_v3/dagger_pulumi_pipeline.py
@@ -9,7 +9,7 @@ from bridge.settings.openedx.types import (
     OpenEdxApplication,
     OpenEdxSupportedRelease,
 )
-from ol_concourse.lib.jobs.infrastructure import pulumi_jobs_chain
+from ol_concourse.pipelines.jobs import pulumi_jobs_chain
 from ol_concourse.lib.models.fragment import PipelineFragment
 from ol_concourse.lib.models.pipeline import (
     AnonymousResource,
