+++ src/ol_concourse/pipelines/open_edx/xqwatcher/docker_pulumi_pipeline.py
@@ -2,7 +2,7 @@ import sys
 
 from bridge.settings.openedx.accessors import filter_deployments_by_application
 from ol_concourse.lib.containers import container_build_task
-from ol_concourse.lib.jobs.infrastructure import pulumi_jobs_chain
+from ol_concourse.pipelines.jobs import pulumi_jobs_chain
 from ol_concourse.lib.models.fragment import PipelineFragment
 from ol_concourse.lib.models.pipeline import (
     GetStep,
